Refactored validateSignup into a module.
