+++ src/utils/index.ts
@@ -1,3 +1,3 @@
 export { getToken, setToken, clearToken } from './tokenHelpers';
-export { validateSignup } from './formValidation';
+export { validateSignup } from './validateSignup';
 export { formatUTCToLocalTimestring } from './time';
