+++ src/utils/validateSignup/index.ts
@@ -1,0 +1,1 @@
+export { default as validateSignup } from './validateSignup';
